Move start reload server into separate module
