+++ lib/server/server.js
@@ -7,7 +7,7 @@
 
 /* eslint-disable no-cond-assign */
 
-function execute(port, options) {
+function execute(port) {
   const extractTranslations = require('../write-translations');
   const metadataUtils = require('./metadataUtils');
   const blog = require('./blog');
@@ -22,9 +22,6 @@ function execute(port, options) {
   const mkdirp = require('mkdirp');
   const glob = require('glob');
   const chalk = require('chalk');
-  const gaze = require('gaze');
-  const tinylr = require('tiny-lr');
-  const constants = require('../core/constants');
   const translate = require('./translate');
   const {renderToStaticMarkupWithDoctype} = require('./renderUtils');
   const feed = require('./feed');
@@ -103,26 +100,6 @@ function execute(port, options) {
         console.error('Request failed:', error);
       }
     });
-  }
-
-  function startLiveReload() {
-    process.env.NODE_ENV = 'development';
-    const server = tinylr();
-    server.listen(constants.LIVE_RELOAD_PORT, () => {
-      console.log(
-        'LiveReload server started on port %d',
-        constants.LIVE_RELOAD_PORT
-      );
-    });
-
-    gaze(
-      [`../${readMetadata.getDocsPath()}/**/*`, '**/*', '!node_modules/**/*'],
-      function() {
-        this.on('all', () => {
-          server.notifyClients(['/']);
-        });
-      }
-    );
   }
 
   reloadMetadata();
@@ -398,7 +375,6 @@ function execute(port, options) {
     requestFile(`http://localhost:${port}${req.path}.html`, res, next);
   });
 
-  if (options.watch) startLiveReload();
   app.listen(port);
 }
