+++ lib/start-server.js
@@ -44,6 +44,11 @@ program
   .option('--no-watch', 'Toggle live reload file watching')
   .parse(process.argv);
 
+function startLiveReloadServer() {
+  const liveReloadServer = require('./server/liveReloadServer.js');
+  liveReloadServer();
+}
+
 let port = parseInt(program.port, 10) || process.env.PORT || 3000;
 let numAttempts = 0;
 const MAX_ATTEMPTS = 10;
@@ -65,9 +70,12 @@ function checkPort() {
         numAttempts += 1;
         checkPort();
       } else {
+        const options = program.opts();
+        if (options.watch) startLiveReloadServer();
+
         // start local server on specified port
         const server = require('./server/server.js');
-        server(port, program.opts());
+        server(port);
         const {baseUrl} = require(`${CWD}/siteConfig.js`);
         const host = `http://localhost:${port}${baseUrl}`;
         console.log('Docusaurus server started on port %d', port);
